Time: 2 ms (31.33%), Space: 59.2 MB (69.62%) - LeetHub

+++ 1343-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold/1343-number-of-sub-arrays-of-size-k-and-average-greater-than-or-equal-to-threshold.cpp
@@ -13,7 +13,7 @@ public:
         while(r < n){
             sum+= arr[r];
 
-            double avg = sum/k;
+            int avg = sum/k;
             if(avg >= threshold)
                 cnt++;
             sum-=arr[l];
